LPS-193057 Inline NullSession into its only usage SharedSessionWrapper as inner class
