--- util-java/src/com/liferay/util/servlet/NullSession.java
@@ -1,153 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.util.servlet;
-
-import com.liferay.petra.string.StringPool;
-import com.liferay.portal.kernel.util.ListUtil;
-import com.liferay.portal.kernel.util.StringUtil;
-
-import java.util.Collections;
-import java.util.Enumeration;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-import javax.servlet.ServletContext;
-import javax.servlet.http.HttpSession;
-import javax.servlet.http.HttpSessionContext;
-
-/**
- * @author Brian Wing Shun Chan
- */
-public class NullSession implements HttpSession {
-
-	public NullSession() {
-		_creationTime = System.currentTimeMillis();
-		_id =
-			NullSession.class.getName() + StringPool.POUND +
-				StringUtil.randomId();
-		_lastAccessedTime = _creationTime;
-		_maxInactiveInterval = 0;
-		_servletContext = null;
-		_new = true;
-	}
-
-	@Override
-	public Object getAttribute(String name) {
-		return _attributes.get(name);
-	}
-
-	@Override
-	public Enumeration<String> getAttributeNames() {
-		return Collections.enumeration(_attributes.keySet());
-	}
-
-	@Override
-	public long getCreationTime() {
-		return _creationTime;
-	}
-
-	@Override
-	public String getId() {
-		return _id;
-	}
-
-	@Override
-	public long getLastAccessedTime() {
-		return _lastAccessedTime;
-	}
-
-	@Override
-	public int getMaxInactiveInterval() {
-		return _maxInactiveInterval;
-	}
-
-	@Override
-	public ServletContext getServletContext() {
-		return _servletContext;
-	}
-
-	/**
-	 * @deprecated As of Bunyan (6.0.x)
-	 */
-	@Deprecated
-	@Override
-	public HttpSessionContext getSessionContext() {
-		return null;
-	}
-
-	/**
-	 * @deprecated As of Wilberforce (7.0.x)
-	 */
-	@Deprecated
-	@Override
-	public Object getValue(String name) {
-		return getAttribute(name);
-	}
-
-	/**
-	 * @deprecated As of Wilberforce (7.0.x)
-	 */
-	@Deprecated
-	@Override
-	public String[] getValueNames() {
-		List<String> names = ListUtil.fromEnumeration(getAttributeNames());
-
-		return names.toArray(new String[0]);
-	}
-
-	@Override
-	public void invalidate() {
-		_attributes.clear();
-	}
-
-	@Override
-	public boolean isNew() {
-		return _new;
-	}
-
-	/**
-	 * @deprecated As of Wilberforce (7.0.x)
-	 */
-	@Deprecated
-	@Override
-	public void putValue(String name, Object value) {
-		setAttribute(name, value);
-	}
-
-	@Override
-	public void removeAttribute(String name) {
-		_attributes.remove(name);
-	}
-
-	/**
-	 * @deprecated As of Wilberforce (7.0.x)
-	 */
-	@Deprecated
-	@Override
-	public void removeValue(String name) {
-		removeAttribute(name);
-	}
-
-	@Override
-	public void setAttribute(String name, Object value) {
-		_attributes.put(name, value);
-	}
-
-	@Override
-	public void setMaxInactiveInterval(int maxInactiveInterval) {
-		_maxInactiveInterval = maxInactiveInterval;
-	}
-
-	private final Map<String, Object> _attributes = new HashMap<>();
-	private final long _creationTime;
-	private final String _id;
-	private final long _lastAccessedTime;
-	private int _maxInactiveInterval;
-	private final boolean _new;
-	private final ServletContext _servletContext;
-
-}
